Thread Version 1.0
Thread Version that works.
However it is currently preforming worse than the unthreaded version. Going to investigate more and see if I can find the issue. Current theory is the thread joining isn't performing correctly, or the overhead time for setup & pre-calculations for threads is just about .8 seconds and that is causing the larger time cost on sequences this small